+++ src/components/common/spinner/spinner.tsx
@@ -1,5 +1,4 @@
 import { CircularProgress } from '@mui/material';
-import React from 'react';
 
 const Spinner = () => {
   return <CircularProgress color="success" />;
